Move example config tests into skypy directory
