--- examples/tests/test_mccl_galaxies.py
@@ -1,9 +1,0 @@
-from skypy.pipeline import load_skypy_yaml, Pipeline
-
-
-def test_mccl_galaxies():
-    config = load_skypy_yaml("../mccl_galaxies.yml")
-    pipeline = Pipeline(config)
-    pipeline.execute()
-
-    assert len(pipeline.state) > 0
